Fix parsing of numbers in exponential notation and negative exponent
A value like 1E-3 results in a float value, even if it not contains the decimal separator. NumberFormatter can parse this correctly, but we need to pass TYPE_DOUBLE. Before it were only used when the string contained the decimal separator, therefore the TYPE_INT was used before (which resulted in the inproper result of 0).