[AssetMapper] Make digest hash algorithm configurable
I've been wanting to use shorter digests in my mapped assets / importmap URL's.

This PR adds an hashAlgorithm argument to the MappedAssetFactory to do so.

I'm not sure this is enough/good choice, as we need some requirement regarding the digests (7+ caracters, alnum only)

But i adding an entire AssetHasherInterface would be probably overkill, so i'm very open for feedback / ideas there.

(I did not put FrameworkBundle modifications in this PR, as the discussion will probably have an impact on what's to do)